refactor: LocationCard into it's own file
